More clang negotiation.

+++ src/ColumnReader.cc
@@ -421,7 +421,7 @@ namespace orc {
       lengthArray[i] += lengthArray[i-1];
     }
     long blobSize = lengthArray[dictionaryCount];
-    dictionaryBlob.resize(blobSize);
+    dictionaryBlob.resize((unsigned long)blobSize);
     std::unique_ptr<SeekableInputStream> blobStream =
       stripe.getStream(columnId, proto::Stream_Kind_DICTIONARY_DATA);
     readFully(dictionaryBlob.data(), blobSize, blobStream.get());
